+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased changes
 
+- `Tablerone.icon` and `Tablerone.path` can take strings, and can get `:otp_app` from opts.
 - Add `Tablerone.path/1` for introspecting the local priv path of svg files.
 - `mix tablerone.download` ensures erlang dependencies are started, sets user agent.